requirements.txt
bcrypt==3.1.7
Click==7.0
Cerberus==1.3.2
dnslib==0.9.10
falcon==2.0.0
gunicorn==20.0.4
peewee==3.13.1
PyJWT==1.7.1
PyMySQL==0.9.3
python-dotenv==0.10.3
redis==3.3.11
shortuuid==0.5.0